+++ pkg/odo/cli/service/create.go
@@ -7,7 +7,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/redhat-developer/odo/pkg/log"
 	scv1beta1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
-	"github.com/redhat-developer/odo/pkg/catalog/ui"
+	"github.com/redhat-developer/odo/pkg/odo/cli/service/ui"
 	"github.com/redhat-developer/odo/pkg/odo/genericclioptions"
 	"github.com/redhat-developer/odo/pkg/odo/util"
 	"github.com/redhat-developer/odo/pkg/odo/util/completion"
